HOCS-6211: Remove integration tests due to CI dependency. Added behavioural tests for correspondent and case-attachment.
