+++ src/test/java/uk/gov/digital/ho/hocs/client/migration/casework/MigrationCaseworkClientTest.java
@@ -9,9 +9,13 @@ import org.springframework.http.ResponseEntity;
 import uk.gov.digital.ho.hocs.application.RestClient;
 import uk.gov.digital.ho.hocs.client.migration.casework.dto.CreateMigrationCaseRequest;
 import uk.gov.digital.ho.hocs.client.migration.casework.dto.CreateMigrationCaseResponse;
+import uk.gov.digital.ho.hocs.client.migration.casework.dto.CreateMigrationCorrespondentRequest;
+import uk.gov.digital.ho.hocs.client.migration.casework.dto.MigrationComplaintCorrespondent;
+import uk.gov.digital.ho.hocs.domain.queue.complaints.CorrespondentType;
 
 import java.time.LocalDate;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -53,5 +57,35 @@ public class MigrationCaseworkClientTest {
         migrationCaseworkClient.migrateCase(request);
 
         verify(restClient).post(serviceUrl, "/migrate/case", request, CreateMigrationCaseResponse.class);
+    }
+
+    @Test
+    public void shouldMigrateCorrespondent() {
+        CreateMigrationCorrespondentRequest request  = new CreateMigrationCorrespondentRequest(
+                UUID.randomUUID(),
+                UUID.randomUUID(),
+                createCorrespondent(),
+                List.of(createCorrespondent())
+        );
+
+        migrationCaseworkClient.migrateCorrespondent(request);
+
+        verify(restClient).post(serviceUrl,"/migrate/correspondent", request, Void.class);
+    }
+
+    private MigrationComplaintCorrespondent createCorrespondent() {
+        return new MigrationComplaintCorrespondent(
+                "fullName",
+                CorrespondentType.COMPLAINANT,
+                "address1",
+                "address2",
+                "address3",
+                "postcode",
+                "country",
+                "organisation",
+                "telephone",
+                "email",
+                "reference"
+        );
     }
 }
